refactor(dir_graph): split insert_node_routed into node_write.rs (god-file gate)
The drift provenance fields tipped dir_graph/mod.rs past the 3000-LoC god-file
ceiling (3012). Per the project rule (split, don't raise the ceiling), extract
the self-contained insert_node_routed (~100 lines, the single node-create
chokepoint where provenance is stamped) into a sibling node_write.rs, mirroring
the existing schema_ops.rs/caches.rs splits. mod.rs back to 2912; no behaviour
or API change.
